+++ task-manager/src/routers/user.js
@@ -12,19 +12,34 @@ router.post('/users', async (req, res) => {
     }
 })
 
+router.post('/users/login', async (req, res,) => {
+    try {
+        const user = await User.findByCredentials(req.body.email, req.body.password)
+        return res.status(200).send(user)
+    } catch (err) {
+        return res.status(400).send(err)
+    }
+})
+
 router.get('/users', async (req, res) => {
     try {
         const users = await User.find({})
-        return res.send(users)
+        if (users.length === 0) return res.status(200).send('No users in dbase')
+        return res.status(200).send(users)
     } catch (err) {
         return res.status(500).send(err)
     }
 })
 
 router.get('/users/:id', async (req, res) => {
-    const _id = req.params.id
     try {
-        const user = await User.findById(_id)
+        if (!req.params.id.match(/^[0-9a-fA-F]{24}$/)) {
+            return res.status(500).send('Wrong ObjectID')
+        }
+        const user = await User.findById(req.params.id)
+        if (!user) {
+            return res.status(400).send('User with id ' + req.params.id + ' not found')
+        }
         return res.send(user)
     } catch (error) {
         return res.status(500).send(error)
@@ -41,7 +56,19 @@ router.patch('/users/:id', async (req, res) => {
         return res.status(400).send({ error : "Can't update"})
     }
     try {
-        return res.status(200).send(await User.findByIdAndUpdate(req.params.id, req.body, {new : true, runValidators: true}))
+        if (!req.params.id.match(/^[0-9a-fA-F]{24}$/)) {
+            return res.status(500).send('Wrong ObjectID')
+        }
+        const user = await User.findById(req.params.id)
+
+        if (!user) {
+            return res.status(400).send('User with id ' + req.params.id + ' not found')
+        }
+        updates.forEach((value) => user[value] = req.body[value])
+
+        return res.status(200).send(await user.save())
+
+        //return res.status(200).send(await User.findByIdAndUpdate(req.params.id, req.body, {new : true, runValidators: true}))
     } catch (err) {
         return res.status(500).send(err)
     }
@@ -49,10 +76,16 @@ router.patch('/users/:id', async (req, res) => {
 
 router.delete('/users/:id', async (req, res) => {
     try {
+        if (!req.params.id.match(/^[0-9a-fA-F]{24}$/)) {
+            return res.status(500).send('Wrong ObjectID')
+        }
         const user = await User.findByIdAndDelete(req.params.id)
+        if (!user) {
+            return res.status(400).send('User with id ' + req.params.id + ' not found')
+        }
         return res.status(200).send(user)
     } catch (error) {
-        res.status(500).send(error)
+        return res.status(500).send(error)
     }
 })
